[04] AGL - High Bill Analyzer/app/requirements.txt
pandas==2.2.2
numpy==1.26.4
fastapi==0.114.0
uvicorn==0.30.1
scikit-learn==1.4.2
jinja2==3.1.4
python-dateutil==2.9.0.post0
matplotlib==3.8.4
